+++ attributes/messaging.rb
@@ -53,7 +53,10 @@ default['openstack']['mq']['auto_delete'] = false
 # Default qpid and rabbit values (for attribute assignment below)
 ###################################################################
 default['openstack']['mq']['qpid']['protocol'] = 'tcp'
+# global switch for handling rabbit ssl
 default['openstack']['mq']['rabbitmq']['use_ssl'] = false
+# global switch for handling rabbit ha
+default['openstack']['mq']['rabbitmq']['ha'] = false
 # defined in oslo/messaging/_drivers/impl_qpid.py
 default['openstack']['mq']['qpid']['topology_version'] = 1
 qpid_defaults = {
@@ -79,7 +82,7 @@ rabbit_defaults = {
   vhost: node['openstack']['mq']['vhost'],
   port: node['openstack']['endpoints']['mq']['port'],
   host: node['openstack']['endpoints']['mq']['host'],
-  ha: false,
+  ha: node['openstack']['mq']['rabbitmq']['ha'],
   use_ssl: node['openstack']['mq']['rabbitmq']['use_ssl']
 }